Show selected POI on map

+++ core/navigation/src/main/java/com/hien/le/dkvfinder/core/navigation/NavigationRoute.kt
@@ -1,6 +1,6 @@
 package com.hien.le.dkvfinder.core.navigation
 
 sealed class NavigationRoute {
-    data class ToMap(val latitude: Float, val longitude: Float) : NavigationRoute()
+    data class ToMap(val latitude: Double, val longitude: Double) : NavigationRoute()
     data class ToPoiDetails(val poiId: Int) : NavigationRoute()
 }
